+++ cmd/sebak/cmd/run.go
@@ -24,7 +24,7 @@ import (
 
 	"boscoin.io/sebak/lib"
 	"boscoin.io/sebak/lib/consensus"
-	"boscoin.io/sebak/lib/node/node_runner"
+	"boscoin.io/sebak/lib/node/noderunner"
 	"strconv"
 )
 
@@ -367,7 +367,7 @@ func runNode() error {
 	// Execution group.
 	var g run.Group
 	{
-		nr, err := node_runner.NewNodeRunner(flagNetworkID, localNode, policy, nt, isaac, st)
+		nr, err := noderunner.NewNodeRunner(flagNetworkID, localNode, policy, nt, isaac, st)
 		conf := &consensus.ISAACConfiguration{
 			TimeoutINIT:       timeoutINIT,
 			TimeoutSIGN:       timeoutSIGN,
