+++ lib/block/init.go
