+++ lib/block/operation.go
