+++ lib/consensus/isaac.go
@@ -5,8 +5,8 @@ import (
 
 	"boscoin.io/sebak/lib/common"
 	"boscoin.io/sebak/lib/node"
-	"boscoin.io/sebak/lib/round"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/consensus/round"
+	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/transaction"
 )
