+++ lib/consensus/running_round.go
@@ -3,8 +3,8 @@ package consensus
 import (
 	"boscoin.io/sebak/lib/common"
 	"boscoin.io/sebak/lib/error"
-	"boscoin.io/sebak/lib/round"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/consensus/round"
+	"boscoin.io/sebak/lib/block"
 )
 
 type RunningRound struct {
