+++ lib/network/api/api_test.go
@@ -12,7 +12,7 @@ import (
 
 	"boscoin.io/sebak/lib/common"
 	"boscoin.io/sebak/lib/storage"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 
 	"boscoin.io/sebak/lib/error"
 	"boscoin.io/sebak/lib/transaction"
