+++ lib/network/api/api_transaction.go
@@ -6,7 +6,7 @@ import (
 	"boscoin.io/sebak/lib/network/httputils"
 	"boscoin.io/sebak/lib/observer"
 
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 	"github.com/gorilla/mux"
 )
