+++ lib/network/api/resource_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"boscoin.io/sebak/lib/storage"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/transaction"
 	"github.com/stretchr/testify/require"
 )
