+++ lib/node/noderunner/node_runner_message_checker_test.go
@@ -1,4 +1,4 @@
-package node_runner
+package noderunner
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 
 	"boscoin.io/sebak/lib/common"
 	"boscoin.io/sebak/lib/error"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/transaction"
 )
