+++ lib/node/noderunner/node_runner_test.go
@@ -1,4 +1,4 @@
-package node_runner
+package noderunner
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 	"boscoin.io/sebak/lib/network"
 	"boscoin.io/sebak/lib/node"
 	"boscoin.io/sebak/lib/storage"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 
 	"boscoin.io/sebak/lib/consensus"
 	"boscoin.io/sebak/lib/transaction"
