+++ lib/node/noderunner/proposer_calculator_test.go
@@ -1,4 +1,4 @@
-package node_runner
+package noderunner
 
 import (
 	"testing"
