+++ lib/storage/statedb/state_object.go
@@ -3,7 +3,7 @@ package statedb
 import (
 	"boscoin.io/sebak/lib/common"
 	"boscoin.io/sebak/lib/observer"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/storage/statedb/trie"
 	"bytes"
 	"fmt"
