+++ lib/storage/statedb/statedb.go
@@ -2,7 +2,7 @@ package statedb
 
 import (
 	"boscoin.io/sebak/lib/common"
-	"boscoin.io/sebak/lib/storage/block"
+	"boscoin.io/sebak/lib/block"
 	"boscoin.io/sebak/lib/storage/statedb/trie"
 	"fmt"
 )
